Refactoring - WIP
For #42
* Delete stats_handler (in favor of result_backend)
* Delete models, put content in oct.result_backends.sqlite
* Remove stats_handler references into HQ
* Remove stats_handler tests
* Remove stats_handler references into results_to_csv, now using sqlite backend
* Remove stats_handler references in run command (replaced with simple function)
* Remove stats_handler references in turret_manager
